Use controllers and SQLAlchemy repositories
